requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# NBA Data APIs
nba-api>=1.4.1
basketball-reference-scraper>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Deep Learning
tensorflow>=2.15.0
keras>=3.0.0

# Model Interpretation
shap>=0.44.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Utilities
joblib>=1.3.0
python-dateutil>=2.8.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter (optional for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
